--- src/student_jekaterina_soldatova/lesson06/level01/level_4_junior/Task_18.txt
@@ -1,5 +1,0 @@
-Создайте класс WhileInfiniteLoop и в нём напишите бесконечный
-цикл при помощи цикла while.
-
-Создайте класс ForInfiniteLoop и в нём напишите бесконечный
-цикл при помощи цикла for.
