--- src/student_jekaterina_soldatova/lesson06/level01/level_4_junior/Task_19.txt
@@ -1,5 +1,0 @@
-Создайте класс WhileInfiniteLoopBreak и в нём продемонстрируйте
-выход из бесконечного цикла реализованного при помощи цикла while.
-
-Создайте класс ForInfiniteLoopBreak и в нём продемонстрируйте
-выход из бесконечного цикла реализованного при помощи цикла for.
