--- src/student_jekaterina_soldatova/lesson06/level01/level_4_junior/Task_20.txt
@@ -1,5 +1,0 @@
-Создайте класс WhileLoopContinue и в нём продемонстрируйте
-как досрочно закончить текущую итерацию цикла while.
-
-Создайте класс ForLoopContinue и в нём продемонстрируйте
-как досрочно закончить текущую итерацию цикла for.
