--- src/student_jekaterina_soldatova/lesson06/level01/level_4_junior/Task_21.txt
@@ -1,11 +1,0 @@
-Создайте класс TwoDimensionalArray и в нём напишите программу в которой:
-- создайте двумерный массив;
-- заполните двумерный массив случайными числами;
-- посчитайте сумму всех чисел в двумерном массиве.
-
-Помните, что если вы видите часть кода, которую можно реализовать
-отдельно от остальной программы, создайте для неё свой метод.
-Воспользуйтесь созданным вами методом в основной программе
-просто вызвав его. Развивайте и тренеруйте умение разбивать
-программу на методы на простых задачах, этот навык вам
-в дальнейшем поможет справиться с более сложными задачами.
