+++ src/student_jekaterina_soldatova/lesson06/level_1_junior/NumberUtils.java
@@ -1,4 +1,4 @@
-package student_jekaterina_soldatova.lesson06.level01;
+package student_jekaterina_soldatova.lesson06.level_1_junior;
 
 class NumberUtils {
