+++ src/student_jekaterina_soldatova/lesson06/level_1_junior/Task_1.java
@@ -1,4 +1,4 @@
-package student_jekaterina_soldatova.lesson06.level01;
+package student_jekaterina_soldatova.lesson06.level_1_junior;
 
 class Task_1 {
     public static void main(String[] args) {
