+++ src/student_jekaterina_soldatova/lesson06/level_1_junior/Task_5.java
@@ -1,4 +1,4 @@
-package student_jekaterina_soldatova.lesson06.level01;
+package student_jekaterina_soldatova.lesson06.level_1_junior;
 
 public class Task_5 {
     public static void main(String[] args) {
