+++ src/student_jekaterina_soldatova/lesson06/level_4_junior/ForInfiniteLoopBreak.java
@@ -1,14 +1,12 @@
-package student_jekaterina_soldatova.lesson06.level01.level_4_junior;
+package student_jekaterina_soldatova.lesson06.level_4_junior;
 
-public class WhileInfiniteLoopBreak {
+public class ForInfiniteLoopBreak {
     public static void main(String[] args) {
-        int i = 1;
-        while (i > 0) {
+        for (int i = 1; i > 0; i++) {
             System.out.println("i = " + i);
             if (i == 5) {
                 break;
             }
-            i++;
         }
     }
 }
