+++ src/student_jekaterina_soldatova/lesson06/level_5_middle/Task_22.txt
@@ -1,0 +1,25 @@
+Пошагово создаём игру крестики нолики.
+
+Дано поле игры крестики нолики размерностью три на три клетки.
+Поле представлено двумерным массивом целых чисел, где:
+-1 - это пустая клетка,
+0 - это клетка в которую сделал ход первый игрок,
+1 - это клетка в которую сделал ход второй игрок.
+
+Создайте класс TicTacToe и в этом классе реализуйте
+метод для определения является ли данная позиция
+победной по горизонталям для указанного игрока.
+
+Метод должен иметь следующую сигнатуру:
+public boolean isWinPositionForHorizontals(int[][] field, int playerToCheck).
+
+Поле передаётся в этот метод как первый параметр int[][] field.
+Знак игрока (0 или 1) передаётся вторым параметром: int playerToCheck.
+
+Метод должет вернуть:
+ - true если в одной из горизонталей есть три подряд клетки
+   занятые знаком указанного игрока (int playerToCheck),
+ - false иначе.
+
+Создать класс TicTacToeTest
+и написать тесты для разработанного метода.
