+++ src/student_jekaterina_soldatova/lesson06/level_5_middle/Task_23.txt
@@ -1,0 +1,18 @@
+Пошагово создаём игру крестики нолики.
+
+В классе TicTacToe реализуйте
+метод для определения является ли данная позиция
+победной по вертикалям для указанного игрока.
+
+Метод должен иметь следующую сигнатуру:
+public boolean isWinPositionForVerticals(int[][] field, int playerToCheck).
+
+Поле передаётся в этот метод как первый параметр int[][] field.
+Знак игрока (0 или 1) передаётся вторым параметром: int playerToCheck.
+
+Метод должет вернуть:
+ - true если в одной из вертикалей есть три подряд клетки
+   занятые знаком указанного игрока (int playerToCheck),
+ - false иначе.
+
+В классе TicTacToeTest написать тесты для разработанного метода.
