+++ src/student_jekaterina_soldatova/lesson06/level_5_middle/Task_25.txt
@@ -1,0 +1,23 @@
+Пошагово создаём игру крестики нолики.
+
+В классе TicTacToe реализуйте
+метод для определения является ли данная позиция
+победной для указанного игрока.
+
+Подсказка: победной позиция является тогда, когда
+позиция победна по горизонталям ИЛИ вертикалям ИЛИ диагоналям.
+Для реализации этого метода воспользуйтесь методами созданными
+ранее. Прочувствуйте как просто собирать решение более сложной
+задачи из небольших методов.
+
+Метод должен иметь следующую сигнатуру:
+public boolean isWinPosition(int[][] field, int playerToCheck).
+
+Поле передаётся в этот метод как первый параметр int[][] field.
+Знак игрока (0 или 1) передаётся вторым параметром: int playerToCheck.
+
+Метод должет вернуть:
+ - true если позиция победна для указанного игрока (int playerToCheck),
+ - false иначе.
+
+В классе TicTacToeTest написать тесты для разработанного метода.
