+++ src/student_jekaterina_soldatova/lesson06/level_5_middle/Task_26.txt
@@ -1,0 +1,19 @@
+Пошагово создаём игру крестики нолики.
+
+В классе TicTacToe реализуйте
+метод для определения является ли данная позиция ничейной.
+
+Подсказка: ничейной позиция является тогда, когда
+позиция не победна ни для одного из игроков и нет ни
+одной пустой клетки (клетки в которой стоит -1).
+
+Метод должен иметь следующую сигнатуру:
+public boolean isDrawPosition(int[][] field).
+
+Поле передаётся в этот метод как первый параметр int[][] field.
+
+Метод должет вернуть:
+ - true если позиция ничейна,
+ - false иначе.
+
+В классе TicTacToeTest написать тесты для разработанного метода.
